requirements.txt
huggingface-hub==0.34.4
faiss-gpu-cu12==1.12.0
fastmcp==2.12.2
sentence-transformers==5.1.0
chonkie==1.2.1